fix comparisons ... problems with macro & statement order
the way i defined the macro with args was wrong. e.g.
COMPARE_BRANCH(a, b, op) {
  bool succeeds = a op b;
  ..
}

if i call like COMPARE_BRANCH(POP(), POP(), >) will expand to
COMPARE_BRANCH(a, b, op) {
  bool succeeds = POP() > POP();
  ..
}
which behaves a little weirdly because if both a and b come from the stack, then the order doesn't matter.